requirements.txt
coverage==4.5.1
Django==2.0.6
django-cors-middleware==1.3.1
django-extensions==2.1.0
djangorestframework==3.8.2
psycopg2-binary==2.7.5
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
